refactor(agent-bff): name the error file after its class
